Beginning the removal of our hard var: hackathon. Was intended originally to be a reusable module but in the compressed time frame that we had... well its this.
My burn down includes the following off the top of my head, generic variable names,
removal of all traces that this was a Hackathon, this entire stupid thing needs
to be modularized. Some of the components are just a brittle, aka TF and Ansible.
Ansible is currently pined to minor v1.3.6 and is in proc of being migrated to 1.4-min

.. sigh